+++ src/Individuals.cpp
@@ -8,9 +8,9 @@
 Individual_row::Individual_row(int c)
 : cols(c) {}
 
-std::vector<double> Individual_row::return_row() {
+std::vector<long double> Individual_row::return_row() {
     double init_val = 0.00; 
-    std::vector<double> vec(this->cols,init_val);
+    std::vector<long double> vec(this->cols,init_val);
     return vec;
 }
 Individual_scores::Individual_scores(int n_w, int n_id)
@@ -20,6 +20,7 @@ Individual_scores::Individual_scores(int n_w, int n_id)
     int idx = 0;
     std::vector<double> foo;
     while(idx < this->num_weights) {
+
         Individual_row row(n_id);
         this->scores.push_back(row.return_row());
         idx++;
