+++ src/Individuals.h
@@ -16,7 +16,7 @@ public:
     std::vector<std::string> IIDs; //temp -- fixme, return with function call
     Individual_scores() = default;
     Individual_scores(int num_weights, int num_ids);
-    std::vector<std::vector<double> > scores;
+    std::vector<std::vector<long double> > scores;
 };
 
 
@@ -25,7 +25,7 @@ class Individual_row
 public:
     int cols;
     Individual_row(int c);
-    std::vector<double> return_row();
+    std::vector<long double> return_row();
 };
 
 #endif /* Individuals_h */
